+++ Changelog.md
@@ -5,3 +5,11 @@
 - Reworked setting input to handle structured setting values (for example stepper motor settings) and updated settings guide
 - Added manager notification of setting changes
 - Added updated Clock.json file for 16x16 led wordsearch
+## Version 1.0.3
+
+- Turned off low power mode in WiFi startup
+- Modified the clock to use a non-blocking version of ntime
+- Update manager now performs manifest download and local version check
+- Added version values to all code file
+- Added alpha display manager
+- Added alpha uart manager